service/flavours/mlflow/requirements_arm.txt
flask
loguru
# kserve

# Modzy Requirements
grpcio-tools
grpcio
grpcio-reflection
protobuf==3.19.4
pyyaml
six
